Touchup on the gc.
Fix a bug that strings were not put to the black list.

I feel this gc is not really doing much due to we don't run
gc_collect() on every step, which seems to be strange.

Also another strangeness is that if multiple tp_obj pointing to
the same gc managed memory end up on the white list, they may get
deleted twice. the *gci flag may ease the tension here, but I
don't see how.