Call convention (#44)
* add arg signatures to func attrs.

* add a test case.

* merge args and func, minor fix on error message.

* Set func attrs in DEF instruction.

* Change the PY calling convention.

Instead of unpacking arguments on the callee side as byte codes,
we now do this on the caller side with C. This also allow us
to properly implement the CPython conflict error if argument is
given both as varargs and kwargs, and also the default value is
evaluated at func define-time instead of call time.

I think it is faster but I don't have a benchmark.

Also fix a few gc error due to forgetting tp_follow.

The new PY calling convention is to pass all arguments on the regs,
followed by varargs and varkw. The callee takes care to
define the local variables for all arguments, varargs, and varkw.